#52 Iron out some reactivity bugs, add missing schedule description to home page
